requirements.txt
flask
flask-marshmallow
marshmallow
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
azure-monitor-opentelemetry-exporter
python-dotenv
